Separate real implementation from test implementation
